+++ php/Terminus/Loggers/Logger.php
@@ -200,6 +200,9 @@ class Logger extends KLogger {
     }
 
     // interpolate replacement values into the message and return
+    if (!is_string($message)) {
+      $message = print_r($message, true);
+    }
     return strtr($message, $replace);
   }