+++ php/class-terminus-command.php
@@ -118,7 +118,7 @@ abstract class TerminusCommand {
     $method = 'GET',
     $options = null
   ) {
-    if (!in_array($realm, array('login', 'user', 'public'))) {
+    if (!in_array($realm, array('auth/refresh', 'login', 'user', 'public'))) {
       Auth::loggedIn();
     }
 
@@ -127,7 +127,7 @@ abstract class TerminusCommand {
 
       if (!in_array($realm, array('login', 'user'))) {
         $options['cookies'] = array(
-          'X-Pantheon-Session' => Session::getValue('session')
+          'X-Pantheon-Session' => Session::getValue('id_token')
         );
         $options['verify']  = false;
       }
@@ -139,9 +139,7 @@ abstract class TerminusCommand {
           'path'  => $path,
         )
       );
-      if (Terminus::getConfig('debug')) {
-        Terminus::log('debug', 'Request URL: ' . $url);
-      }
+      Terminus::getLogger()->debug('Request URL: {url}', compact('url'));
       $resp = Request::send($url, $method, $options);
       $json = $resp->getBody(true);
 
@@ -201,9 +199,9 @@ abstract class TerminusCommand {
 
     $url = 'https://' . TERMINUS_HOST . '/api/' . $path;
 
-    if (Session::getValue('session')) {
+    if (Session::getValue('id_token')) {
       $req_options['cookies'] = array(
-      'X-Pantheon-Session' => Session::getValue('session')
+      'X-Pantheon-Session' => Session::getValue('id_token')
       );
       $req_options['verify']  = false;
     }